+++ structurizr-component/src/test/java/com/structurizr/component/supporting/implementation/ExampleRepository.java
@@ -1,0 +1,4 @@
+package com.structurizr.component.supporting.implementation;
+
+public interface ExampleRepository {
+}