+++ structurizr-component/src/test/java/com/structurizr/component/supporting/implementation/ExampleRepositoryImpl.java
@@ -1,0 +1,4 @@
+package com.structurizr.component.supporting.implementation;
+
+public class ExampleRepositoryImpl implements ExampleRepository {
+}